convert radio controls to TS
